requirements.txt
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.0
google-genai==1.12.1
flask==3.1.0